+++ test/test_variable.py
@@ -143,6 +143,16 @@ def test_variable_sum(x):
     assert res.nterm == 10
 
 
+def test_variable_sum_warn_using_dims(x):
+    with pytest.warns(DeprecationWarning):
+        x.sum(dims="first")
+
+
+def test_variable_sum_warn_unknown_kwargs(x):
+    with pytest.raises(ValueError):
+        x.sum(unknown_kwarg="first")
+
+
 def test_variable_where(x):
     x = x.where([True] * 4 + [False] * 6)
     assert isinstance(x, linopy.variables.Variable)