+++ linopy/expressions.py
@@ -33,6 +33,7 @@ from linopy.common import (
     forward_as_properties,
     generate_indices_for_printout,
     get_index_map,
+    has_optimized_model,
     print_single_expression,
     to_dataframe,
 )
@@ -48,6 +49,16 @@ from linopy.constants import (
     TERM_DIM,
 )
 
+SUPPORTED_CONSTANT_TYPES = (
+    np.number,
+    int,
+    float,
+    DataArray,
+    pd.Series,
+    pd.DataFrame,
+    np.ndarray,
+)
+
 
 def exprwrap(method, *default_args, **new_default_kwargs):
     @functools.wraps(method)
@@ -58,7 +69,9 @@ def exprwrap(method, *default_args, **new_default_kwargs):
             method(_expr_unwrap(expr), *default_args, *args, **kwargs), expr.model
         )
 
-    _exprwrap.__doc__ = f"Wrapper for the xarray {method} function for linopy.Variable"
+    _exprwrap.__doc__ = (
+        f"Wrapper for the xarray {method.__qualname__} function for linopy.Variable"
+    )
     if new_default_kwargs:
         _exprwrap.__doc__ += f" with default arguments: {new_default_kwargs}"
 
@@ -159,6 +172,7 @@ class LinearExpressionGroupby:
 
             int_map = None
             if isinstance(group, pd.DataFrame):
+                group = group.reindex(self.data.indexes[group.index.name])
                 int_map = get_index_map(*group.values.T)
                 orig_group = group
                 group = group.apply(tuple, axis=1).map(int_map)
@@ -279,14 +293,16 @@ class LinearExpression:
         if data is None:
             da = xr.DataArray([], dims=[TERM_DIM])
             data = Dataset({"coeffs": da, "vars": da, "const": 0.0})
-        elif isinstance(data, DataArray):
-            # assume only constant are passed
-            const = fill_missing_coords(data)
+        elif isinstance(data, SUPPORTED_CONSTANT_TYPES):
+            const = as_dataarray(data)
             da = xr.DataArray([], dims=[TERM_DIM])
             data = Dataset({"coeffs": da, "vars": da, "const": const})
         elif not isinstance(data, Dataset):
+            supported_types = ", ".join(
+                map(lambda s: s.__qualname__, (*SUPPORTED_CONSTANT_TYPES, Dataset))
+            )
             raise ValueError(
-                f"data must be an instance of xarray.Dataset or xarray.DataArray, got {type(data)}"
+                f"data must be an instance of {supported_types}, got {type(data)}"
             )
 
         if not set(data).issuperset({"coeffs", "vars"}):
@@ -337,7 +353,7 @@ class LinearExpression:
         ndim = len(dims)
         dim_sizes = list(self.coord_sizes.values())
         size = np.prod(dim_sizes)  # that the number of theoretical printouts
-        masked_entries = self.mask.sum().values if self.mask is not None else 0
+        masked_entries = (~self.mask).sum().values if self.mask is not None else 0
         lines = []
 
         header_string = self.type
@@ -397,10 +413,14 @@ class LinearExpression:
     def __add__(self, other):
         """
         Add an expression to others.
+
+        Note: If other is a numpy array or pandas object without axes names,
+        dimension names of self will be filled in other
         """
-        other = as_expression(
-            other, model=self.model, coords=self.coords, dims=self.coord_dims
-        )
+        if np.isscalar(other):
+            return self.assign(const=self.const + other)
+
+        other = as_expression(other, model=self.model, dims=self.coord_dims)
         return merge(self, other, cls=self.__class__)
 
     def __radd__(self, other):
@@ -410,10 +430,14 @@ class LinearExpression:
     def __sub__(self, other):
         """
         Subtract others from expression.
+
+        Note: If other is a numpy array or pandas object without axes names,
+        dimension names of self will be filled in other
         """
-        other = as_expression(
-            other, model=self.model, coords=self.coords, dims=self.coord_dims
-        )
+        if np.isscalar(other):
+            return self.assign(const=self.const - other)
+
+        other = as_expression(other, model=self.model, dims=self.coord_dims)
         return merge(self, -other, cls=self.__class__)
 
     def __neg__(self):
@@ -447,6 +471,14 @@ class LinearExpression:
             assert set(coeffs.shape) == set(self.coeffs.shape)
             const = self.const * multiplier
             return self.assign(coeffs=coeffs, const=const)
+
+    def __pow__(self, other):
+        """
+        Power of the expression with a coefficient. The only coefficient allowed is 2.
+        """
+        if not other == 2:
+            raise ValueError("Power must be 2.")
+        return self * self
 
     def __rmul__(self, other):
         """
@@ -496,6 +528,30 @@ class LinearExpression:
         raise NotImplementedError(
             "Inequalities only ever defined for >= rather than >."
         )
+
+    def add(self, other):
+        """
+        Add an expression to others.
+        """
+        return self.__add__(other)
+
+    def sub(self, other):
+        """
+        Subtract others from expression.
+        """
+        return self.__sub__(other)
+
+    def mul(self, other):
+        """
+        Multiply the expr by a factor.
+        """
+        return self.__mul__(other)
+
+    def div(self, other):
+        """
+        Divide the expr by a factor.
+        """
+        return self.__div__(other)
 
     @property
     def loc(self):
@@ -559,6 +615,30 @@ class LinearExpression:
     @property
     def mask(self):
         return None
+
+    @has_optimized_model
+    def _map_solution(self):
+        """
+        Replace variable labels by solution values.
+        """
+        m = self.model
+        sol = pd.Series(m.matrices.sol, m.matrices.vlabels)
+        sol.loc[-1] = np.nan
+        idx = np.ravel(self.vars)
+        values = sol[idx].values.reshape(self.vars.shape)
+        return xr.DataArray(values, dims=self.vars.dims, coords=self.vars.coords)
+
+    @property
+    def solution(self):
+        """
+        Get the optimal values of the expression.
+
+        The function raises an error in case no model is set as a
+        reference or the model is not optimized.
+        """
+        vals = self._map_solution()
+        sol = (self.coeffs * vals).sum(TERM_DIM) + self.const
+        return sol.rename("solution")
 
     @classmethod
     def _sum(cls, expr: Union["LinearExpression", Dataset], dims=None) -> Dataset:
@@ -1203,10 +1283,14 @@ class QuadraticExpression(LinearExpression):
     def __add__(self, other):
         """
         Add an expression to others.
+
+        Note: If other is a numpy array or pandas object without axes names,
+        dimension names of self will be filled in other
         """
-        other = as_expression(
-            other, model=self.model, coords=self.coords, dims=self.coord_dims
-        )
+        if np.isscalar(other):
+            return self.assign(const=self.const + other)
+
+        other = as_expression(other, model=self.model, dims=self.coord_dims)
         if type(other) is LinearExpression:
             other = other.to_quadexpr()
         return merge(self, other, cls=self.__class__)
@@ -1226,10 +1310,14 @@ class QuadraticExpression(LinearExpression):
     def __sub__(self, other):
         """
         Subtract others from expression.
+
+        Note: If other is a numpy array or pandas object without axes names,
+        dimension names of self will be filled in other
         """
-        other = as_expression(
-            other, model=self.model, coords=self.coords, dims=self.coord_dims
-        )
+        if np.isscalar(other):
+            return self.assign(const=self.const - other)
+
+        other = as_expression(other, model=self.model, dims=self.coord_dims)
         if type(other) is LinearExpression:
             other = other.to_quadexpr()
         return merge(self, -other, cls=self.__class__)
@@ -1243,6 +1331,18 @@ class QuadraticExpression(LinearExpression):
             return other.__sub__(self)
         else:
             NotImplemented
+
+    @property
+    def solution(self):
+        """
+        Get the optimal values of the expression.
+
+        The function raises an error in case no model is set as a
+        reference or the model is not optimized.
+        """
+        vals = self._map_solution()
+        sol = (self.coeffs * vals.prod(FACTOR_DIM)).sum(TERM_DIM) + self.const
+        return sol.rename("solution")
 
     @classmethod
     def _sum(cls, expr: "QuadraticExpression", dims=None) -> Dataset: