+++ doc/release_notes.rst
@@ -8,6 +8,10 @@ Upcoming Version
 
 * The LinearExpression and QuadraticExpression class have a new attribute `solution` which returns the optimal values of the expression if the underlying model was optimized.
 
+**Deprecations**
+
+* The argument `dims` in the `.sum` function of variables and expressions was deprecated in favor of the `dim` argument. This aligns the argument name with the xarray convention.
+
 
 Version 0.3.6
 -------------